+++ express-zod-api/src/common-helpers.ts
@@ -83,17 +83,19 @@ export const getMessageFromError = (error: Error): string => {
 /** Takes the original unvalidated examples from the properties of ZodObject schema shape */
 export const pullExampleProps = <T extends z.SomeZodObject>(subject: T) =>
   Object.entries(subject.shape).reduce<Partial<z.input<T>>[]>(
-    (acc, [key, schema]) =>
-      combinations(
+    (acc, [key, schema]) => {
+      const { _def } = schema as z.ZodType;
+      return combinations(
         acc,
-        (schema._def[metaSymbol]?.examples || []).map(objOf(key)),
+        (_def[metaSymbol]?.examples || []).map(objOf(key)),
         ([left, right]) => ({ ...left, ...right }),
-      ),
+      );
+    },
     [],
   );
 
 export const getExamples = <
-  T extends z.ZodTypeAny,
+  T extends z.ZodType,
   V extends "original" | "parsed" | undefined,
 >({
   schema,
@@ -180,6 +182,3 @@ export const isProduction = memoizeWith(
   () => process.env.TSUP_STATIC as string, // eslint-disable-line no-restricted-syntax -- substituted by TSUP
   () => process.env.NODE_ENV === "production", // eslint-disable-line no-restricted-syntax -- memoized
 );
-
-export const nonEmpty = <T>(subject: T[] | ReadonlyArray<T>) =>
-  subject.length ? subject : undefined;