+++ express-zod-api/src/metadata.ts
@@ -9,6 +9,7 @@ export interface Metadata {
   /** @override ZodDefault::_def.defaultValue() in depictDefault */
   defaultLabel?: string;
   brand?: string | number | symbol;
+  isDeprecated?: boolean;
 }
 
 /** @link https://github.com/colinhacks/zod/blob/3e4f71e857e75da722bd7e735b6d657a70682df2/src/types.ts#L485 */
@@ -19,15 +20,15 @@ export const cloneSchema = <T extends z.ZodType>(schema: T) => {
   return copy;
 };
 
-export const copyMeta = <A extends z.ZodTypeAny, B extends z.ZodTypeAny>(
+export const copyMeta = <A extends z.ZodType, B extends z.ZodType>(
   src: A,
   dest: B,
 ): B => {
-  if (!(metaSymbol in src._def)) return dest;
-  const result = cloneSchema(dest);
-  result._def[metaSymbol].examples = combinations(
-    result._def[metaSymbol].examples,
-    src._def[metaSymbol].examples,
+  if (!(metaSymbol in src._def)) return dest; // ensure metadata in src below
+  const result = cloneSchema(dest); // ensures metadata in result below
+  result._def[metaSymbol]!.examples = combinations(
+    result._def[metaSymbol]!.examples,
+    src._def[metaSymbol]!.examples,
     ([destExample, srcExample]) =>
       typeof destExample === "object" && typeof srcExample === "object"
         ? mergeDeepRight({ ...destExample }, { ...srcExample })